Add files via upload
Throughout this notebook, we will practice structuring data in Python. Before getting started, watch the associated instructional video and complete the in-video question. All of the code we will be implementing and related instructions are contained in this notebook. It contains functions like merge, groupby, to_datetime, sort, and isocalendar,